Handle nginx sigquit and container restart #30 (#31)

+++ docker-entrypoint.sh
@@ -3,7 +3,9 @@
 
 set -e
 
-export GITHUB_BASIC_AUTH=$(echo "${GITHUB_USER}:${GITHUB_ACCESS_TOKEN}" | base64)
-envsubst '${GITHUB_BASIC_AUTH}' < /etc/nginx/conf.d/community.deepfence.io.conf.template > /etc/nginx/conf.d/community.deepfence.io.conf
-envsubst '${GITHUB_BASIC_AUTH}' < /etc/nginx/conf.d/docs.deepfence.io.conf.template > /etc/nginx/conf.d/docs.deepfence.io.conf
-rm -f /etc/nginx/conf.d/community.deepfence.io.conf.template /etc/nginx/conf.d/docs.deepfence.io.conf.template
+if [ -f "/etc/nginx/conf.d/community.deepfence.io.conf.template" ]; then
+  export GITHUB_BASIC_AUTH=$(echo "${GITHUB_USER}:${GITHUB_ACCESS_TOKEN}" | base64)
+  envsubst '${GITHUB_BASIC_AUTH}' </etc/nginx/conf.d/community.deepfence.io.conf.template >/etc/nginx/conf.d/community.deepfence.io.conf
+  envsubst '${GITHUB_BASIC_AUTH}' </etc/nginx/conf.d/docs.deepfence.io.conf.template >/etc/nginx/conf.d/docs.deepfence.io.conf
+  rm -f /etc/nginx/conf.d/community.deepfence.io.conf.template /etc/nginx/conf.d/docs.deepfence.io.conf.template
+fi
